Move and Rename custom propType definitions in LayoutPropTypes
